Refactored Clayton meta data extensions
Moved it all into a single file and added additional tests for Clayton
