--- copulae/copula/extensions/__init__.py
@@ -1,2 +1,0 @@
-from .clayton import ClaytonExt
-from .generate import generate_meta
